# apiAulaFinal2.0/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Musicas",
      "version": "1.0.0",
      "description": "API para cadastro de Musicas"
    },
    "paths": {
      "/musicas": {
        "get": {
          "summary": "Obter a lista de musicas",
          "responses": {
            "200": {
              "description": "Lista de musicas retornada com sucesso"
            }
          }
        },
        "post": {
          "summary": "Cadastrar uma nova musica",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/musicas"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Musica adicionada com sucesso"
            },
            "400": {
              "description": "A musica não foi adicionada"
            }
          }
        }
      },
      "/musicas/{id}": {
        "put": {
          "summary": "Atualizar informações da musica por id(O Projeto não tem a função de puxar por id)",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/musicas"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Informações da musica atualizadas com sucesso"
            },
            "404": {
              "description": "Musica não encontrada"
            }
          }
        },
        "delete": {
          "summary": "Excluir uma musica pelo ID(Eata função não foi aplicada)",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Musica excluído com sucesso"
            },
            "404": {
              "description": "Musica não encontrado"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "musicas": {
          "type": "object",
          "properties": {
            "nome": {
              "type": "string"
            },
            "artista": {
              "type": "string"
            },
            "genero": {
                "type": "string"
            },
            "imagem": {
                "type": "string"
            },
            "id": {
                "type": "integer"
            }
          }
        }
      }
    }
  }